+++ metadata.rb
@@ -4,10 +4,10 @@ maintainer_email 'package@datadoghq.com'
 license          'Apache-2.0'
 description      'Installs/Configures datadog components'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '2.13.0'
-chef_version     '>= 12.7' if respond_to? :chef_version
-source_url       'https://github.com/DataDog/chef-datadog' if respond_to? :source_url
-issues_url       'https://github.com/DataDog/chef-datadog/issues' if respond_to? :issues_url
+version          '2.16.1'
+chef_version     '>= 12.7'
+source_url       'https://github.com/DataDog/chef-datadog'
+issues_url       'https://github.com/DataDog/chef-datadog/issues'
 
 %w[
   amazon
@@ -22,9 +22,9 @@ issues_url       'https://github.com/DataDog/chef-datadog/issues' if respond_to?
   supports os
 end
 
-depends          'apt' # We recommend '>= 2.1.0'. See CHANGELOG.md for details
-depends          'chef_handler', '>= 1.1' # We recommend '~> 1.3' with Chef < 12. See CHANGELOG.md for details
-depends          'yum', '>= 3.0' # Use '~> 3.0' with Chef < 12
+depends 'apt' # We recommend '>= 2.1.0'. See CHANGELOG.md for details
+depends 'chef_handler', '>= 1.1' # We recommend '~> 1.3' with Chef < 12. See CHANGELOG.md for details
+depends 'yum', '>= 3.0' # Use '~> 3.0' with Chef < 12
 
 recipe 'datadog::default', 'Default'
 recipe 'datadog::dd-agent', 'Installs the Datadog Agent'